PhoneApp: New in call proximity sensor option
Background: AOSP default behavior is to turn off the proximity sensor when the user have a wired or bluetooth headset connected to the device.

Added option to keep the proximity sensor on during calls using a headset. The option is uncheckd by default, preserving AOSP behavior.

Navigate to Phone -> Settings and check the option Keep proximity sensor on.

Commit includes option in english and translation to portuguese.

Change-Id: I5b09c425f40f918fa616a1c45e6765d4f5c239ab

Conflicts:

	res/values-pt/strings.xml

Signed-off-by: iceandfire <arhamjamal@gmail.com>